Don't nest songs using their beatmapset id.
Because when you import an v7 or an earlier osu file format version, BeatmapSet ID and Beatmap ID will not be in the osu file. Therefore beatmapset_id will be null.

id is randomly generated using uuidv4

user_id is removed because the user page can still be visited using only the name.