Observer pattern + Library
Player and Preview System are designed using an observer pattern

To receive updates from helpers, components needs to have a useState created. Using useEffect one needs to call addObserver passing the setter of the useState on mount. Or removeObserver on unmount. The helper will use the setter to update the State.

Downloaded songs are also now shown on the library view and songs can be played.